+++ docs/CHANGES.md
@@ -1,8 +1,8 @@
 2017-06-20  Release 1.82.5
 
   * Bug fix: Update deprecated SonarQube API for fetching a component.
-  * Bug fix: New API for fetching component version number doesn't work before SonarQube v6.3, fallback on older
-    API if necessary.
+  * Bug fix: New API for fetching component version number and last analyse date and time doesn't work 
+    before SonarQube v6.3, fallback on older API if necessary.
 
 
 2017-06-16  Release 1.82.4